+++ test/fixtures/unless.rb
@@ -20,3 +20,7 @@ end
 foo = unless bar then baz end
 -
 foo = (baz unless bar)
+%
+unless foo += 1
+  foo
+end