+++ test/fixtures/while.rb
@@ -15,3 +15,7 @@ end
 foo = while bar do baz end
 -
 foo = (baz while bar)
+%
+while foo += 1
+  foo
+end